requirements.txt
appdirs==1.4.3
appnope==0.1.0
click==6.7
decorator==4.0.11
Flask==0.12.1
Flask-REST-JSONAPI==0.12.2
Flask-Uploads==0.2.1
Flask-Webpack==0.1.0
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==2.13.1
marshmallow-jsonapi==0.11.1
packaging==16.8
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.1
